requirements.txt
openai==0.27.1
requests==2.28.2
termcolor==2.2.0
ipykernel
textract==1.6.5
llama-index==0.6.24
langchain==0.0.239
python-dotenv==1.0.0
httpx
transformers
faiss-cpu==1.7.2
six==1.12.0
typing-extensions==4.5.0
urllib3<2
unstructured[pdf]
langchain_community
selenium